+++ dotProduct.h
@@ -9,7 +9,7 @@
 #define MAGNITUDE 10 //17*15*15
 #define GRIDPOINTS 10
 
-void dotProd(double A[MAGNITUDE][GRIDPOINTS], double B[MAGNITUDE], double out[MAGNITUDE]);
+void dotProd(int A[MAGNITUDE][GRIDPOINTS], int B[MAGNITUDE], int out[MAGNITUDE]);
 
 
 #endif //FWI_DOTPRODUCT_H
